fix bug cho quản lý người dùng

+++ GUI/fQuanLyNguoiDung.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
+using System.Diagnostics.Eventing.Reader;
 using System.Drawing;
 using System.Linq;
 using System.Text;
@@ -24,11 +25,12 @@ namespace GUI
         private void fQuanLyNguoiDung_Load(object sender, EventArgs e)
         {
             txtMaNguoiDung.Enabled = false;
-            txtTenDangNhap.Enabled = false; 
+            txtTenDangNhap.Enabled = false;
             dgvNguoiDung.DataSource = NguoiDungBUS.Instance.GetAllNguoiDung();
             cboNhomNguoiDung.DataSource = NhomNguoiDungBUS.Instance.GetAllNhomNguoiDung();
             cboNhomNguoiDung.DisplayMember = "TENNHOM";
             cboNhomNguoiDung.ValueMember = "ID";
+            cboNhomNguoiDung.Text = "";
             txtMaNhomNguoiDung.Enabled = false;
             dgvNhomNguoiDung.DataSource = NhomNguoiDungBUS.Instance.GetAllNhomNguoiDung();
             dgvNhomNguoiDung.Columns["ID"].Visible = false;
@@ -56,17 +58,50 @@ namespace GUI
         }
         private void btnSuaNguoiDung_Click(object sender, EventArgs e)
         {
-            NguoiDung nguoidung = new NguoiDung();
-            nguoidung.MaNguoiDung = txtMaNguoiDung.Text;
-            nguoidung.TenNguoiDung = txtTenNguoiDung.Text;
-            nguoidung.IDNhomNguoiDung = Convert.ToInt32(cboNhomNguoiDung.SelectedValue.ToString());
-            NguoiDungBUS.Instance.UpdateNguoiDung(nguoidung);
-            dgvNguoiDung.DataSource = NguoiDungBUS.Instance.GetAllNguoiDung();
+            if(txtMaNguoiDung.Text == "")
+            {
+                MessageBox.Show("Vui lòng chọn thông tin trong datagridview để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }    
+            else if(txtTenNguoiDung.Text == "")
+            {
+                MessageBox.Show("Vui lòng nhập tên người dùng!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                txtTenNguoiDung.Focus();
+            }
+            else if (cboNhomNguoiDung.SelectedValue != null)
+            {
+                if (MessageBox.Show("Bạn thực sự có muốn thay đổi thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+                {
+                    NguoiDung nguoidung = new NguoiDung();
+                    nguoidung.MaNguoiDung = txtMaNguoiDung.Text;
+                    nguoidung.TenNguoiDung = txtTenNguoiDung.Text;
+                    nguoidung.IDNhomNguoiDung = Convert.ToInt32(cboNhomNguoiDung.SelectedValue.ToString());
+                    NguoiDungBUS.Instance.UpdateNguoiDung(nguoidung);
+                    dgvNguoiDung.DataSource = NguoiDungBUS.Instance.GetAllNguoiDung();
+                    MessageBox.Show("Sửa thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+            }
+            else
+            {
+                MessageBox.Show("Vui lòng chọn lại nhóm người dùng!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboNhomNguoiDung.Focus();
+            }    
         }
         private void btnXoaNguoiDung_Click(object sender, EventArgs e)
         {
-            NguoiDungBUS.Instance.DeleteNguoiDung(txtMaNguoiDung.Text);
-            dgvNguoiDung.DataSource = NguoiDungBUS.Instance.GetAllNguoiDung();
+            if (txtMaNguoiDung.Text == "")
+            {
+                MessageBox.Show("Vui lòng chọn thông tin trong datagridview để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (MessageBox.Show("Bạn thực sự có muốn xoá thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+            {
+                NguoiDungBUS.Instance.DeleteNguoiDung(txtMaNguoiDung.Text);
+                dgvNguoiDung.DataSource = NguoiDungBUS.Instance.GetAllNguoiDung();
+                MessageBox.Show("Xoá thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                txtMaNguoiDung.Text = "";
+                txtTenNguoiDung.Text = "";
+                txtTenDangNhap.Text = "";
+                cboNhomNguoiDung.Text = "";
+            }
         }
 
         private void btnThemDocGia_Click(object sender, EventArgs e)
@@ -85,17 +120,45 @@ namespace GUI
 
         private void btnSuaNhomNguoiDung_Click(object sender, EventArgs e)
         {
-            NhomNguoiDung NND = new NhomNguoiDung();
-            NND.MaNhom = txtMaNhomNguoiDung.Text;
-            NND.TenNhom = txtTenNhomNguoiDung.Text;
-            NhomNguoiDungBUS.Instance.UpdateNhomNguoiDung(NND);
-            dgvNhomNguoiDung.DataSource = NhomNguoiDungBUS.Instance.GetAllNhomNguoiDung();
+            if (txtMaNhomNguoiDung.Text == "")
+            {
+                MessageBox.Show("Vui lòng chọn thông tin trong datagridview để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (txtTenNhomNguoiDung.Text == "")
+            {
+                MessageBox.Show("Vui lòng nhập tên nhóm người dùng!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                txtTenNhomNguoiDung.Focus();
+            }
+            else if (NhomNguoiDungBUS.Instance.GetNhomNguoiDungByTenNhom(txtTenNhomNguoiDung.Text).Rows.Count > 0 )
+            {
+                MessageBox.Show("Nhóm đã tồn tại. Vui lòng nhập lại tên nhóm!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                txtTenNhomNguoiDung.Focus();
+            }    
+            else if (MessageBox.Show("Bạn thực sự có muốn thay đổi thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+            {
+                NhomNguoiDung NND = new NhomNguoiDung();
+                NND.MaNhom = txtMaNhomNguoiDung.Text;
+                NND.TenNhom = txtTenNhomNguoiDung.Text;
+                NhomNguoiDungBUS.Instance.UpdateNhomNguoiDung(NND);
+                dgvNhomNguoiDung.DataSource = NhomNguoiDungBUS.Instance.GetAllNhomNguoiDung();
+                MessageBox.Show("Sửa thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+            } 
         }
 
         private void btnXoaNhomNguoiDung_Click(object sender, EventArgs e)
         {
-            NhomNguoiDungBUS.Instance.DeleteNhomNguoiDung(txtMaNhomNguoiDung.Text);
-            dgvNhomNguoiDung.DataSource = NhomNguoiDungBUS.Instance.GetAllNhomNguoiDung();
+            if (txtMaNhomNguoiDung.Text == "")
+            {
+                MessageBox.Show("Vui lòng chọn thông tin trong datagridview để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (MessageBox.Show("Bạn thực sự có muốn xoá thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+            {
+                NhomNguoiDungBUS.Instance.DeleteNhomNguoiDung(txtMaNhomNguoiDung.Text);
+                dgvNhomNguoiDung.DataSource = NhomNguoiDungBUS.Instance.GetAllNhomNguoiDung();
+                MessageBox.Show("Xoá thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                txtMaNhomNguoiDung.Text = "";
+                txtTenNhomNguoiDung.Text = "";
+            }
         }
 
         private void btnThemNhomNguoiDung_Click(object sender, EventArgs e)
@@ -115,18 +178,55 @@ namespace GUI
 
         private void btnSuaChucNang_Click(object sender, EventArgs e)
         {
-            ChucNang chucnang = new ChucNang();
-            chucnang.MaChucNang = txtMaChucNang.Text;
-            chucnang.TenChucNang = txtTenChucNang.Text;
-            chucnang.TenManHinhDuocLoad = txtTenManHinhDuocLoad.Text;
-            ChucNangBUS.Instance.UpdateChucNang(chucnang);
-            dgvChucNang.DataSource = ChucNangBUS.Instance.GetAllChucNang();
+            if (txtMaChucNang.Text == "")
+            {
+                MessageBox.Show("Vui lòng chọn thông tin trong datagridview để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (txtTenChucNang.Text == "")
+            {
+                MessageBox.Show("Vui lòng nhập tên chức năng!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                txtTenChucNang.Focus();
+            }
+            else if (txtTenManHinhDuocLoad.Text == "")
+            {
+                MessageBox.Show("Vui lòng nhập tên màn hình được load!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                txtTenManHinhDuocLoad.Focus();
+            }
+            else
+            {
+                ChucNang chucnang = new ChucNang();
+                chucnang.MaChucNang = txtMaChucNang.Text;
+                chucnang.TenChucNang = txtTenChucNang.Text;
+                chucnang.TenManHinhDuocLoad = txtTenManHinhDuocLoad.Text;
+                if (ChucNangBUS.Instance.GetChucNangByTenChucNang(chucnang).Rows.Count > 0)
+                {
+                    MessageBox.Show("Chức năng đã tồn tại. Vui lòng nhập chức năng khác!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    txtTenChucNang.Focus();
+                }
+                else if (MessageBox.Show("Bạn thực sự có muốn thay đổi thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+                {
+                    ChucNangBUS.Instance.UpdateChucNang(chucnang);
+                    dgvChucNang.DataSource = ChucNangBUS.Instance.GetAllChucNang();
+                    MessageBox.Show("Sửa thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+            }
         }
 
         private void btnXoaChucNang_Click(object sender, EventArgs e)
         {
-            ChucNangBUS.Instance.DeleteChucNang(txtMaChucNang.Text);
-            dgvChucNang.DataSource = ChucNangBUS.Instance.GetAllChucNang();
+            if (txtMaChucNang.Text == "")
+            {
+                MessageBox.Show("Vui lòng chọn thông tin trong datagridview để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+            }
+            else if (MessageBox.Show("Bạn thực sự có muốn xoá thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+            {
+                ChucNangBUS.Instance.DeleteChucNang(txtMaChucNang.Text);
+                dgvChucNang.DataSource = ChucNangBUS.Instance.GetAllChucNang();
+                MessageBox.Show("Xoá thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                txtMaChucNang.Text = "";
+                txtTenChucNang.Text = "";
+                txtTenManHinhDuocLoad.Text = "";
+            }    
         }
 
         private void btnThemChucNang_Click(object sender, EventArgs e)
@@ -145,6 +245,9 @@ namespace GUI
                     int ID = Convert.ToInt32(cboMaNhomNguoiDungPQ.SelectedValue.ToString());
                     txtTenNhomNguoiDungPQ.Text = NhomNguoiDungBUS.Instance.GetNhomNguoiDungByID(ID).Rows[0][2].ToString();
                     dgvPhanQuyen.DataSource = PhanQuyenBUS.Instance.GetAllPhanQuyenByIDNhomNguoiDung(ID);
+                    cboChucNangPQ.Text = "";
+                    txtTenChucNangPQ.Text = "";
+                    txtTenManHinhDuocLoadPQ.Text = "";
                 }
                 catch { }
             }
@@ -164,26 +267,72 @@ namespace GUI
 
         private void btnXoaPhanQuyen_Click(object sender, EventArgs e)
         {
-            PhanQuyen phanquyen = new PhanQuyen();
-            phanquyen.IDNhomNguoiDung = Convert.ToInt32(cboMaNhomNguoiDungPQ.SelectedValue.ToString());
-            phanquyen.IDChucNang = Convert.ToInt32(cboChucNangPQ.SelectedValue.ToString());
-            PhanQuyenBUS.Instance.DeletePhanQuyen(phanquyen);
-            dgvPhanQuyen.DataSource = PhanQuyenBUS.Instance.GetAllPhanQuyenByIDNhomNguoiDung(phanquyen.IDNhomNguoiDung);
-            cboChucNangPQ.Text = "";
-            txtTenChucNangPQ.Text = "";
-            txtTenManHinhDuocLoadPQ.Text = "";
+            if (cboMaNhomNguoiDungPQ.SelectedValue == null)
+            {
+                MessageBox.Show("Vui lòng chọn mã nhóm để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboMaNhomNguoiDungPQ.Focus();
+            }
+            else if (cboChucNangPQ.SelectedValue == null)
+            {
+                MessageBox.Show("Vui lòng chọn mã chức năng để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboChucNangPQ.Focus();
+            }
+            else
+            {
+                PhanQuyen phanquyen = new PhanQuyen();
+                phanquyen.IDNhomNguoiDung = Convert.ToInt32(cboMaNhomNguoiDungPQ.SelectedValue.ToString());
+                phanquyen.IDChucNang = Convert.ToInt32(cboChucNangPQ.SelectedValue.ToString());
+                if (PhanQuyenBUS.Instance.GetAllPhanQuyenByPhanQuyen(phanquyen).Rows.Count > 0)
+                {
+                    if (MessageBox.Show("Bạn thực sự có muốn xoá thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+                    {
+                        PhanQuyenBUS.Instance.DeletePhanQuyen(phanquyen);
+                        dgvPhanQuyen.DataSource = PhanQuyenBUS.Instance.GetAllPhanQuyenByIDNhomNguoiDung(phanquyen.IDNhomNguoiDung);
+                        MessageBox.Show("Xoá thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                        cboChucNangPQ.Text = "";
+                        txtTenChucNangPQ.Text = "";
+                        txtTenManHinhDuocLoadPQ.Text = "";
+                    }
+                }
+                else 
+                {
+                    MessageBox.Show("Chức năng không tồn tại. Vui lòng chọn lại mã chức năng để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    cboChucNangPQ.Focus();
+                }
+            }      
         }
 
         private void btnThemPhanQuyen_Click(object sender, EventArgs e)
         {
-            PhanQuyen phanquyen = new PhanQuyen();
-            phanquyen.IDNhomNguoiDung = Convert.ToInt32(cboMaNhomNguoiDungPQ.SelectedValue.ToString());
-            phanquyen.IDChucNang = Convert.ToInt32(cboChucNangPQ.SelectedValue.ToString());
-            PhanQuyenBUS.Instance.InsertPhanQuyen(phanquyen);
-            dgvPhanQuyen.DataSource = PhanQuyenBUS.Instance.GetAllPhanQuyenByIDNhomNguoiDung(phanquyen.IDNhomNguoiDung);
-            cboChucNangPQ.Text = "";
-            txtTenChucNangPQ.Text = "";
-            txtTenManHinhDuocLoadPQ.Text = "";
+            if (cboMaNhomNguoiDungPQ.SelectedValue == null)
+            {
+                MessageBox.Show("Vui lòng chọn mã nhóm để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboMaNhomNguoiDungPQ.Focus();
+            }
+            else if (cboChucNangPQ.SelectedValue == null)
+            {
+                MessageBox.Show("Vui lòng chọn mã chức năng để thực hiện!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                cboChucNangPQ.Focus();
+            }
+            else
+            {
+                PhanQuyen phanquyen = new PhanQuyen();
+                phanquyen.IDNhomNguoiDung = Convert.ToInt32(cboMaNhomNguoiDungPQ.SelectedValue.ToString());
+                phanquyen.IDChucNang = Convert.ToInt32(cboChucNangPQ.SelectedValue.ToString());
+                if (PhanQuyenBUS.Instance.GetAllPhanQuyenByPhanQuyen(phanquyen).Rows.Count > 0)
+                {
+                    MessageBox.Show("Chức năng đã tồn tại. Vui lòng nhập chức năng khác!", "Lỗi!", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    cboChucNangPQ.Focus();
+                }
+                else if (MessageBox.Show("Bạn thực sự có muốn thay đổi thông tin!", "Thông tin!", MessageBoxButtons.OKCancel, MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.OK)
+                {
+                    PhanQuyenBUS.Instance.InsertPhanQuyen(phanquyen);
+                    dgvPhanQuyen.DataSource = PhanQuyenBUS.Instance.GetAllPhanQuyenByIDNhomNguoiDung(phanquyen.IDNhomNguoiDung);
+                    MessageBox.Show("Thêm thông tin thành công!", "Thông tin", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                
+            }    
+            
         }
 
         private void cboChucNangPQ_SelectedIndexChanged(object sender, EventArgs e)
@@ -200,6 +349,7 @@ namespace GUI
             }
             else
             {
+                cboChucNangPQ.Text = "";
                 txtTenChucNangPQ.Text = "";
                 txtTenManHinhDuocLoadPQ.Text = "";
             }
